feat: Add local development environment variables
The commit adds new environment variables for local development:
`URL_APP`, which points to `http://localhost:3000`, and `NODE_ENV`,
which is set to `development`. These variables enable the application to
run and behave differently in a local development environment. This
change improves the development experience by providing the necessary
configuration options while working on the project locally.

Issue #123